Added the compute_all Function and Structured Project Added the compute_all Function to find all possible substrings between likely matches.  The project is also now structured to separate frontend, backend, and testing directories
