+++ backend/file_to_fingerprint.py
@@ -1,4 +1,4 @@
-class filetofingerprint():
+class FileToFingerprint():
     def __init__(self, filename, fileid, keys, similarto):
         self.filename = filename
         self.fileid = fileid
